+++ tests/Integration/RestAPI/Emulator/RestAPI.spec.ts
@@ -6,6 +6,7 @@ import { opts, RUN_INTEGRATION_TESTS } from "../helper";
 const api = new Api(opts);
 
 describe('RestAPI - Emulator', () => {
+	let emulatorUid: string;
 	before(function () {
 		// in order to run these tests, fill in auth and RUN_INTEGRATION_TESTS environment variables (please see '../helper.ts' file)
 		if (!RUN_INTEGRATION_TESTS || (opts.accountAuth as any).tokenId === '' || (opts.accountAuth as any).token === '') {
@@ -23,8 +24,14 @@ describe('RestAPI - Emulator', () => {
 		const [emulator] = await api.emulator.list();
 		should(emulator instanceof Emulator).be.true();
 		should(emulator.uid).be.not.empty();
+		emulatorUid = emulator.uid;
 		should(emulator.duid).be.not.empty();
 		should(emulator.name).be.not.empty();
 		should(emulator.createdAt instanceof Date).be.true();
+	});
+
+	it('should delete old emulator', async () => {
+		await api.emulator.delete(emulatorUid);
+		should(true).be.true();
 	});
 });