+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added method to get list of emulators for given account and method to create new emulator
+- Added method to delete old emulator
 
 ## [1.4.2] - 2020-10-30
 ### Deprecated